improve club lookup
